Improve index use when scanning aggregated billing records.
Getting the DB settings and creating the DB engine need to be separated to avoid a
circular import.

It's not clear how big the performance of the index will be as I don't have enough
data for PostgreSQL to even try, but with enable_seqscan=0 the index is used and is
in the correct order for the GROUP BY (eliminating a sort).

Ultimately we may need to aggregate older billing events into day-sized ones in the
DB, but this may be enough for now.
